requirements.txt
fastapi
uvicorn
python-multipart
openai
pydantic
python-dotenv
pytest
httpx
